restructured and renamed external map kernels
