+++ src/runtime/local/kernels/MAP_EXTERNAL/CtypesMapKernel_sharedMem_voidPointer.py
