--- src/runtime/local/kernels/MAP_NUMPY/CMakeLists.txt
@@ -1,7 +1,0 @@
-set (CMAKE_INCLUDE_CURRENT_DIR ON)
-set(CMAKE_CXX_STANDARD 17 CACHE STRING "C++ standard to conform to")
-set(CMAKE_CXX_STANDARD_REQUIRED ON)
-cmake_minimum_required(VERSION 3.21)
-
-project(mapExternalPL_NumpyPythonKernel)
-message(STATUS "NumpyPythonKernel CMake Build")
