+++ test/CMakeLists.txt
@@ -72,7 +72,6 @@ set(TEST_SOURCES
         runtime/local/kernels/AggColTest.cpp
         runtime/local/kernels/AggCumTest.cpp
         runtime/local/kernels/AggRowTest.cpp
-        runtime/local/kernels/AlternativeMapTest.cpp
         runtime/local/kernels/CartesianTest.cpp
         runtime/local/kernels/CastObjTest.cpp
         runtime/local/kernels/CastObjScaTest.cpp
@@ -101,6 +100,7 @@ set(TEST_SOURCES
         runtime/local/kernels/IsSymmetricTest.cpp
         runtime/local/kernels/NumDistinctApproxTest.cpp
         runtime/local/kernels/MapTest.cpp
+        runtime/local/kernels/AlternativeMapTest.cpp
         runtime/local/kernels/MatMulTest.cpp
         runtime/local/kernels/OrderTest.cpp
         runtime/local/kernels/OuterBinaryTest.cpp
